refactor: share category + AL display helpers across tables
Pure code motion, no behaviour change:
- getCategoryColor: StudentTable.tsx -> lib/student-utils.ts
- AL display order/rank/short-label/chip-colour: CumulativeALTable.tsx
  -> new lib/academic-level.ts

Both are about to gain a second consumer (the Student Results table), and a
duplicated copy would let the Enrollment roster and the Performance tab drift
on what "SC" or "M1" looks like.
